+++ CHANGELOG.md
@@ -10,9 +10,14 @@ Note that Sockeye has checks in place to not translate with an old model that wa
 
 Each version section may have have subsections for: _Added_, _Changed_, _Removed_, _Deprecated_, and _Fixed_.
 
+## [1.18.54]
+### Added
+- `--source-factor-vocabs` can be set to provide source factor vocabularies.
+
 ## [1.18.53]
 ### Added
-- `--source-factor-vocabs` can be set for source factor vocabularies.
+- Always skipping softmax for greedy decoding by default, only for single models.
+- Added option `--skip-topk` for greedy decoding.
 
 ## [1.18.52]
 ### Fixed